refactor(preview): split 1019-line Preview.tsx into focused modules
Preview.tsx had grown to 1019 lines with Header / TabBar / Body /
ErrorBody / LinesBody / XmlBody / MediaBody / HalfBlockImage / a
dozen helpers / the main effect all in one file. Reading it cold
required scrolling past 600 lines of helpers before the component
even started.

Splits the file by responsibility — each named subcomponent gets
its own file, type and helper definitions move to peer modules:

  body/Header.tsx          path + mode badge row
  body/TabBar.tsx          internal tab strip
  body/Body.tsx            Switch dispatcher over ContentState
  body/MediaBody.tsx       metadata card + image surface
  body/HalfBlockImage.tsx  RGB grid -> half-block char rows
  body/LinesBody.tsx       file / diff text rendering
  body/XmlBody.tsx         SVG syntax-highlight rendering
  body/ErrorBody.tsx       error + hint copy

  content-state.ts   ContentState + MediaContent type union
  format.ts          describeMediaKind + formatMtime + formatBytes re-export
  error-summary.ts   summarizePreviewError + looksBinary
  image-budget.ts    terminal-size-driven render budget

Preview.tsx is now 427 lines and reads as a shell — props, state
signals, the (path, mode) effect that routes to loadContent, the
key-binding registration, and the JSX tree. Behavior unchanged:
all 98 preview tests still pass; typecheck and biome clean.

No public API change. Barrel still re-exports Preview, PreviewProps,
PreviewApi from ./Preview.
